requirements.txt
fastapi
uvicorn
python-chess
sse-starlette